+++ test/consistency-and-style/test_summary_text.py
@@ -10,7 +10,6 @@ from mathics import __file__ as mathics_initfile_path
 from mathics.builtin import name_is_builtin_symbol
 from mathics.builtin.base import Builtin
 from mathics.doc.common_doc import skip_doc
-from mathics.version import __version__  # noqa used in loading to check consistency.
 
 # Get file system path name for mathics.builtin
 mathics_path = osp.dirname(mathics_initfile_path)
